move svg to correct folder
