+++ apps/studio/src/components/Svg/NoSearchResultSvgr.tsx
